\name{COVID19_2020}
\alias{COVID19_2020}
\docType{data}
\title{
COVID-19 case data for 138 countries and 18 variables
}
\description{
Sample datasets for Generalized Network-based Dimensionality Reduction and Analysis (GNDA)

Covid'19 of countries (2020), where the data frame has 138 observations of 18 variables.
}
\usage{data("COVID19_2020")}
\format{
  A data frame with 138 observations 18 variables.
}

\source{
Kurbucz, M. T. (2020). A joint dataset of official COVID-19 reports and the governance, trade and competitiveness indicators of World Bank group platforms. Data in brief, 31, 105881.
}

\examples{
data(COVID19_2020)

}
\keyword{datasets}
