\name{CrimesUSA1990.X}
\alias{CrimesUSA1990.X}
\docType{data}
\title{
Crimes in USA cities in 1990. Independent variables (X)
}
\description{
Sample datasets for Generalized Network-based Dimensionality Reduction and Analysis (GNDA)

Crimes in USA cities in 1990. Independent variables (X)
}
\usage{data("CrimesUSA1990.X")}
\format{
  A data frame with 1994 observations 123 variables.
}

\source{
UCI - Machine Learning Repository: https://archive.ics.uci.edu/ml/datasets/communities+and+crime
}

\examples{
data(CrimesUSA1990.X)

}
\keyword{datasets}
