\name{CrimesUSA1990.Y}
\alias{CrimesUSA1990.Y}
\docType{data}
\title{
Crimes in USA cities in 1990. Dependent variable (Y)
}
\description{
Sample datasets for Network-based Dimensionality Reduction and Analysis (NDA)

Crimes in USA cities in 1990. Dependent variable (Y)
}
\usage{data("CrimesUSA1990.Y")}
\format{
  A data frame with 1994 observations 1 variables.
}

\source{
UCI - Machine Learning Repository: https://archive.ics.uci.edu/ml/datasets/communities+and+crime
}

\examples{
data(CrimesUSA1990.Y)

}
\keyword{datasets}
