\name{CWTS_2020}
\alias{CWTS_2020}
\docType{data}
\title{
CWTS Leiden University Ranking 2020 with 1,176 universities and 42 indicators
}
\description{
Sample datasets for Generalized Network-based Dimensionality Reduction and Analysis (GNDA)

CWTS Leiden's 2020 dataset, where the data frame has 1176 observations of 42 variables.
}
\usage{data("CWTS_2020")}
\format{
  A data frame with 1176 observations of 42 variables.
}

\source{
CWTS Leiden Ranking 2020: https://www.leidenranking.com/ranking/2020/list
}

\examples{
data(CWTS_2020)

}
\keyword{datasets}
