\name{GOVDB2020}
\alias{GOVDB2020}
\docType{data}
\title{
Governmental and economic data for 138 countries and 2,161 variables
}
\description{
Sample datasets for Generalized Network-based Dimensionality Reduction and Analysis (GNDA)

Governmental and economic data of countries (2020), where the data frame has 138 observations of 2161 variables.
}
\usage{data("GOVDB2020")}
\format{
  A data frame with 138 observations of 2161 variables.
}

\source{
Kurbucz, M. T. (2020). A joint dataset of official COVID-19 reports and the governance, trade and competitiveness indicators of World Bank group platforms. Data in brief, 31, 105881.
}

\examples{
data(GOVDB2020)

}
\keyword{datasets}
