\name{nda-package}
\alias{nda-package}
\alias{nda}
\docType{package}

\title{Package of Generalized Network-based Dimensionality Reduction and Analyses}

\description{
The package of Generalized Network-based Dimensionality Reduction and Analysis (GNDA).
}

\seealso{\code{\link{ndr}}, \code{\link{ndrlm}}, \code{\link{plot}}, \code{\link{biplot}}, \code{\link{summary}}, \code{\link{dCor}}.}

\author{Zsolt T. Kosztyan*, Marcell T. Kurbucz, Attila I. Katona, Zahid Khan

e-mail*: kosztyan.zsolt@gtk.uni-pannon.hu}

\references{
Kosztyan, Z. T., Kurbucz, M. T., & Katona, A. I. (2022). Network-based dimensionality reduction of high-dimensional, low-sample-size datasets. Knowledge-Based Systems, 109180.

Kosztyan, Z. T., Katona, A. I., Kurbucz, M. T., & Lantos, Z. (2024). Generalized network-based dimensionality analysis. Expert Systems with Applications, 238, 121779. <URL:
https://doi.org/10.1016/j.eswa.2023.121779>.
}

\keyword{package}
