\name{I40_2020}
\alias{I40_2020}
\docType{data}
\title{
NUTS2 regional development data (2020) of I4.0 readiness, where the data frame has 414 observations of 101 variables.
}
\description{
Sample datasets for Generalized Network-based Dimensionality Reduction and Analysis (GNDA)

NUTS2 regional development data (2020), where the data frame has 414 observations of 101 variables.
}
\usage{data("COVID19_2020")}
\format{
  A data frame with 414 observations of 101 variables.
}

\source{
Honti, G., Czvetkó, T., & Abonyi, J. (2020). Data describing the regional Industry 4.0 readiness index. Data in Brief, 33, 106464.
}

\examples{
data(I40_2020)

}
\keyword{datasets}
