\name{nda}
\alias{nda}
\docType{package}

\title{Package of Generalized Network-based Dimensionality Reduction and Analyses}

\description{
The package of Generalized Network-based Dimensionality Reduction and Analysis (GNDA).
}

\seealso{\code{\link{ndr}}, \code{\link{plot}}, \code{\link{biplot}}, \code{\link{summary}}, \code{\link{dCor}}.}

\author{Zsolt T. Kosztyan*, Marcell T. Kurbucz, Attila I. Katona

e-mail*: kosztyan.zsolt@gtk.uni-pannon.hu}

\references{
Kosztyan, Z. T., Kurbucz, M. T., & Katona, A. I. (2022). Network-based dimensionality reduction of high-dimensional, low-sample-size datasets. Knowledge-Based Systems, 109180.

Kurbucz, M. T., Katona, A. I., Lantos, Z., & Kosztyan, Z. T. (2021). The Role of Societal Aspects in the Formation of Official COVID-19 Reports: A Data-Driven Analysis. International journal of environmental research and public health, 18(4), 1505.
}

\keyword{package}
