\name{nda}
\alias{nda}
\docType{package}

\title{Package of Network-based Dimensionality reduction and Analysis}

\description{
Network-based Dimensionality reduction and Analysis. Biplot function.
}

\seealso{\code{\link{ndr}}, \code{\link{plot}}, \code{\link{biplot}}, \code{\link{summary}}, \code{\link{dCor}}.}

\author{Zsolt T. Kosztyan*, Marcell T. Kurbucz, Attila I. Katona

e-mail*: kzst@gtk.uni-pannon.hu}

\references{
Kurbucz, M. T., Katona, A. I., Lantos, Z., & Kosztyan, Z. T. (2021). The Role of Societal Aspects in the Formation of Official COVID-19 Reports: A Data-Driven Analysis. International journal of environmental research and public health, 18(4), 1505.
}

\keyword{package}
